Split out common metadata validation code
